refactor : insert 메서드 위치 변경(CourtService -> NewCourtService)
