--- src/main/java/org/slams/server/court/exception/UserNotFoundException.java
@@ -1,10 +1,0 @@
-package org.slams.server.court.exception;
-
-import org.slams.server.common.error.exception.EntityNotFoundException;
-import org.slams.server.common.error.exception.ErrorCode;
-
-public class UserNotFoundException extends EntityNotFoundException {
-    public UserNotFoundException(String message) {
-        super(message);
-    }
-}
